+++ changelog.md
@@ -58,6 +58,7 @@ This version is not yet released. If you are reading this on the website, then t
 ### Website
 - Hide experimental glyphs in the editor by default
   - They can be toggled on in the settings
+- An `# Experimental!` comment can now be easily inserted via a settings button or with `Ctrl+E`
 
 ## 0.9.5 - 2024-02-28
 ### Interpreter